erstattet lokale variable med size_t og intptr_t

+++ Lib/Src/Util/PropertyContainer.cpp
@@ -2,7 +2,7 @@
 #include <PropertyChangeListener.h>
 
 void PropertyContainer::notifyPropertyChanged(int id, const void *oldValue, const void *newValue) const {
-  for(int i = 0; i < (int)m_listeners.size(); i++) {
+  for(size_t i = 0; i < m_listeners.size(); i++) {
     m_listeners[i]->handlePropertyChanged(this, id, oldValue, newValue);
   }
 }
@@ -16,7 +16,7 @@ void PropertyContainer::addPropertyChangeListener(PropertyChangeListener *listen
 }
 
 void PropertyContainer::removePropertyChangeListener(PropertyChangeListener *listener) {
-  const int index = (int)m_listeners.getFirstIndex(listener);
+  const intptr_t index = m_listeners.getFirstIndex(listener);
   if(index >= 0) {
     m_listeners.remove(index);
   }
